requirements.txt
alembic==1.0.11
autopep8==1.4.4
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
Click==7.0
cloudinary==1.17.0
dnspython==1.16.0
Flask==1.1.1
Flask-Assets==0.12
Flask-CKEditor==0.4.3
Flask-Cors==3.0.8
Flask-Dropzone==1.5.4
Flask-FontAwesome==0.1.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
mailer==0.8.1
Mako==1.1.0
MarkupSafe==1.1.1
mock==3.0.5
mongoengine==0.18.2
pycodestyle==2.5.0
pycparser==2.19
pymongo==3.9.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.7
urllib3==1.25.3
webassets==0.12.1
Werkzeug==0.15.5
WTForms==2.2.1
